refactor: home MenuPage の同居サブコンポーネントを ui/ へ分解し命名を可読化 (FRESTYLE-210)
MenuPage.tsx(305行) に同居していた 4 コンポーネント(と型・定数)を FSD の
page-local な ui/ へ抽出。振る舞いは不変。

- ui/: FeatureSection / FeatureCard(+CardColor/iconBg) / MenuSkeleton / StatsSkeleton
- 省略名を改名: s→sectionIndex / i→cardIndex・index / t→tech
- 不要になった Link・ArrowRightIcon・LanguageIcon の import を FeatureCard 側へ移動

MenuPage は 305→178 行。既存テスト(MenuPage.test.tsx 9本ほか)は assert 不変で
全通過(DOM 出力不変を担保)。tsc / vitest(1221) / eslint / build /
branches 80.16%(閾値維持) すべて green。
